paws: add kubernetes basic ingress setup
This is the kubernetes setup for the ingress. Is mostly copied over from Toolforge with a few
changes, the most important one being that we specify a toleration to only run nginx-ingress in the
ingress worker nodes.

I also dropped the tool-fourohfour setup from Toolforge, and other minor changes.

Eventually, we will probably add Ingress objects configuration in the same yaml file.

This patch also adds the PSP yaml configuration file.

Bug: T195217
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>