URL encode path to file after replacing user name with user id
Make URLs safe for processing by nginx before returning from LUA by
reapplying URL encoding following replacement of the `User:<name>` path
component with the associated Wikimedia SUL account's uid. See
<https://github.com/openresty/lua-nginx-module#ngxescape_uri> for the
docs on using `ngx.escape_uri`. We need the escape to run as a full URI
escape in this case so that whitespace is encoded, but path separators
are not.

Bug: T258304